requirements.txt
fastapi
uvicorn
google-generativeai
duckduckgo-search
slowapi
python-jose
pydantic-settings
python-multipart
requests
python-dotenv
